+++ inst/include/dplyr/Collecter.h
@@ -624,7 +624,7 @@ inline Collecter* collecter(SEXP model, int n) {
     break;
   }
 
-  stop("Unsupported vector type %s", Rf_type2char(TYPEOF(model)));
+  stop("unsupported vector type %s", Rf_type2char(TYPEOF(model)));
 }
 
 inline Collecter* promote_collecter(SEXP model, int n, Collecter* previous) {
@@ -666,7 +666,7 @@ inline Collecter* promote_collecter(SEXP model, int n, Collecter* previous) {
   default:
     break;
   }
-  stop("Unsupported vector type %s", Rf_type2char(TYPEOF(model)));
+  stop("unsupported vector type %s", Rf_type2char(TYPEOF(model)));
 }
 
 }