+++ inst/include/dplyr/OrderVisitorImpl.h
@@ -9,6 +9,7 @@
 #include <dplyr/OrderVisitor.h>
 #include <dplyr/DataFrameVisitors.h>
 #include <dplyr/MatrixColumnVisitor.h>
+#include <dplyr/bad.h>
 
 namespace dplyr {
 
@@ -192,33 +193,38 @@ private:
 };
 
 
-inline OrderVisitor* order_visitor(SEXP vec, bool ascending);
+inline OrderVisitor* order_visitor(SEXP vec, const SymbolString& name, const bool ascending, const int i);
 
 template <bool ascending>
-OrderVisitor* order_visitor_asc(SEXP vec);
+OrderVisitor* order_visitor_asc(SEXP vec, const SymbolString& name);
 
 template <bool ascending>
 OrderVisitor* order_visitor_asc_matrix(SEXP vec);
 
 template <bool ascending>
-OrderVisitor* order_visitor_asc_vector(SEXP vec);
+OrderVisitor* order_visitor_asc_vector(SEXP vec, const SymbolString& name);
 
-inline OrderVisitor* order_visitor(SEXP vec, bool ascending) {
-  if (ascending) {
-    return order_visitor_asc<true>(vec);
+inline OrderVisitor* order_visitor(SEXP vec, const SymbolString& name, const bool ascending, const int i) {
+  try {
+    if (ascending) {
+      return order_visitor_asc<true>(vec, name);
+    }
+    else {
+      return order_visitor_asc<false>(vec, name);
+    }
   }
-  else {
-    return order_visitor_asc<false>(vec);
+  catch (const Rcpp::exception& e) {
+    bad_pos_arg(i + 1, e.what());
   }
 }
 
 template <bool ascending>
-inline OrderVisitor* order_visitor_asc(SEXP vec) {
+inline OrderVisitor* order_visitor_asc(SEXP vec, const SymbolString& name) {
   if (Rf_isMatrix(vec)) {
     return order_visitor_asc_matrix<ascending>(vec);
   }
   else {
-    return order_visitor_asc_vector<ascending>(vec);
+    return order_visitor_asc_vector<ascending>(vec, name);
   }
 }
 
@@ -236,7 +242,7 @@ inline OrderVisitor* order_visitor_asc_matrix(SEXP vec) {
   case DPLYR_CPLXSXP:
     return new OrderVisitorMatrix<CPLXSXP, ascending>(vec);
   case DPLYR_VECSXP:
-    stop("Matrix can't be a list", Rf_type2char(TYPEOF(vec)));
+    stop("Matrix can't be a list");
   }
 
   stop("Unreachable");
@@ -244,7 +250,7 @@ inline OrderVisitor* order_visitor_asc_matrix(SEXP vec) {
 }
 
 template <bool ascending>
-inline OrderVisitor* order_visitor_asc_vector(SEXP vec) {
+inline OrderVisitor* order_visitor_asc_vector(SEXP vec, const SymbolString& name) {
   switch (TYPEOF(vec)) {
   case INTSXP:
     return new OrderVectorVisitorImpl<INTSXP, ascending, Vector<INTSXP > >(vec);
@@ -267,7 +273,7 @@ inline OrderVisitor* order_visitor_asc_vector(SEXP vec) {
     break;
   }
 
-  stop("Unsupported vector type %s", Rf_type2char(TYPEOF(vec)));
+  stop("unsupported vector type %s", Rf_type2char(TYPEOF(vec)));
 }
 }